refactor: split bot.go into separate handler files
Extract commands, handlers, and helpers into dedicated files
for better cohesion and maintainability.
