+++ bot/handler_hello.go
@@ -1,0 +1,12 @@
+package bot
+
+import "github.com/bwmarrin/discordgo"
+
+func handleHello(s *discordgo.Session, i *discordgo.InteractionCreate) {
+	s.InteractionRespond(i.Interaction, &discordgo.InteractionResponse{
+		Type: discordgo.InteractionResponseChannelMessageWithSource,
+		Data: &discordgo.InteractionResponseData{
+			Content: "Hello Command! 😃",
+		},
+	})
+}
